fix: setup job cleanup sidecar issue
This fix the sidecar issues in jobs. Because a sidecar is not stopped
when the job finished and there is no good solution to stop a container
in a pod. That's why we delete the job after it's finished when sidecars
are used. Note migrations are currently not deleted... This must be done
differently.