fix: Timing checks on job completed
In local env it's working sometimes but not always, so we remove this
timing check and always return a running if the container name is not
found. This issue can only be created by the operator itself and since
we have e2e tests we catch this now by and endless running state for
jobs.