feat: add more visibility for deployment updates
This includes scaling from other resources, we now go back to
initilization when the scaling is changed by the operator but manuell
changes are only putting the store resource for a short time on init
state. This is because available replicas != replicas for the
deployment. But the show desired replicas is from the store resource,
that's why they can be different and it's fine.